refactor(tests): consolidate OSS and Auth route security tests into 1:1 test files
