+++ tests/test_init.py
@@ -49,10 +49,6 @@ ALLOWLIST_NON_CONFORMING_TESTS = {
     "server/mcp/test_as_discovery.py",
     "server/mcp/test_client.py",
     "server/mcp/test_oauth_flow.py",
-    "server/routes/test_auth_rate_limit.py",
-    "server/routes/test_oss_security.py",
-    "server/routes/test_oss_upload_security.py",
-    "server/routes/test_oss_upload_security_extra.py",
     "server/routes/test_system_extended.py",
     "server/services/processor_modules/test_processor_integration.py",
     "server/services/processor_modules/test_summary_module.py",
